+++ src/pages/Quiz.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Question from '../components/Question';
 import Loading from '../pages/Loading';
 
-import { pickRandom } from './helpers';
+import { pickRandom } from '../helpers';
 
 type CMQuestion = {
   id: number;
